requirement.txt
flask
google-generativeai
weaviate-client